+++ src/lib/heifs.ts
@@ -6,11 +6,12 @@ import utilsFactory, { handleLibraryError, isStandardError } from './utils'
 
 type ResponseBody = {
   created: string[];
+  error?: { message: string };
 }
 
-type ErrorOptionalMessage = ResponseBody & { error?: { message: string } }
+type ErrorOptionalMessage = ResponseBody
 const errorSchema = (message: string): ErrorOptionalMessage => {
-  const out = { created: [] }
+  const out: ResponseBody = { created: [] }
   if (!message) return out
   return { ...out, error: { message } }
 }
@@ -24,7 +25,6 @@ function uniqueHeifs(files: Filesystem[]) {
   const groupedFiles = files.reduce((groups: Record<string, Filesystem[]>, file) => {
     const nameWithoutExt = basename(file.name, file.ext)
     if (!groups[nameWithoutExt]) {
-
       groups[nameWithoutExt] = []
     }
     groups[nameWithoutExt].push(file)
@@ -36,7 +36,7 @@ function uniqueHeifs(files: Filesystem[]) {
       filteredFiles.some((file) => file.ext.toLocaleLowerCase() === 'heic')
       && !filteredFiles.some((file) => file.ext.toLocaleLowerCase() === 'jpg'),
     )
-    .flat()
+    .flatMap((group) => group.filter((file) => file.ext === 'heic'))
 
   return heifFilesWithoutJpg
 }
@@ -79,7 +79,6 @@ async function post(
   try {
     const heifs: string[] = []
     for (const file of uniqueHeifs(files)) {
-
       heifs.push(await processHeif(file, destinationPath))
     }